+++ packages/gateway/src/server/control-api/resolvers/index.ts
@@ -1,0 +1,7 @@
+import updateApi from './update-api';
+
+export default {
+  Query: {
+    updateApi,
+  },
+};