+++ packages/types/src/api-def-config.ts
@@ -5,6 +5,7 @@ export interface ApiDefConfig {
   endpoint: string;
   source_config_names?: string[];
   schema_polling_interval?: number;
+  schema_updates_through_control_api?: boolean;
   mesh?: {
     serve?: ServeConfig;
     require?: string[];